Removendo erros

+++ src/main/java/com/cgp/banco/dao/GenericDAO.java
@@ -3,9 +3,9 @@ package com.cgp.banco.dao;
 import java.util.List;
 
 public interface GenericDAO<T, ID> {
-    void setUserId(Integer userId);
+    void setUserId(Long userId);
 
-    Integer getUserId();
+    Long getUserId();
 
     // Salva uma nova entidade no banco de dados
     void salvar(T entity);
